backend/pyproject.toml
[project]
name = "bla"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.121.1",
    "pillow>=12.0.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "qrcode>=8.2",
    "sqlalchemy>=2.0.44",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
]
